+++ packages/ui5-test-writer/src/index.ts
@@ -1,2 +1,3 @@
 export { generateOPAFiles } from './fiori-elements-opa-writer';
 export { generateFreestyleOPAFiles } from './fiori-freestyle-opa-writer';
+export { addVirtualTestConfig } from './utils/opaQUnitUtils';