+++ packages/app-config-writer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/app-config-writer
 
+## 0.7.1
+
+### Patch Changes
+
+- 2f0c182: feat: support virtual preview endpoints for test generation
+
+    When `useVirtualPreviewEndpoints` is enabled, test harness files (testsuite, unitTests, opaTests) are served virtually and not written to disk. UI5 yaml files are updated with `flp.path: test/flp.html` and test framework entries (OPA5, Testsuite, QUnit) are added to ui5-mock.yaml.
+
 ## 0.7.0
 
 ### Minor Changes